refactor : 스토리북과 컴포넌트를 가깝게 위치
1. preview.ts에 globalStyle 설정
2. 스토리북과 컴포넌트를 가깝게 위치하도록 폴더 수정
