+++ .storybook/preview.ts
@@ -1,4 +1,6 @@
 import type { Preview } from "@storybook/react";
+import GlobalStyles from "../src/GlobalStyles";
+import { withThemeFromJSXProvider } from "@storybook/addon-themes";
 
 const preview: Preview = {
   parameters: {
@@ -35,5 +37,11 @@ const preview: Preview = {
     },
   },
 };
+
+export const decorators = [
+  withThemeFromJSXProvider({
+    GlobalStyles,
+  }),
+];
 
 export default preview;
